Fix installation step.

+++ src/uIntra.Notification/Installer/Migrations/NotificationSettingsInstallationStep_0_2_4_6.cs
@@ -167,6 +167,7 @@ namespace uIntra.Notification.Installer.Migrations
         {
             var mailTemplateFolderXpath = XPathHelper.GetXpath(_documentTypeAliasProvider.GetDataFolder(), _documentTypeAliasProvider.GetMailTemplateFolder());
             var publishedContent = _umbracoHelper.TypedContentSingleAtXPath(mailTemplateFolderXpath);
+            if (publishedContent is null) return;
             var content = contentService.GetById(publishedContent.Id);
             contentService.Delete(content);
         }
